Added support for fluctuating class names
New configurable properties `fluctuateListing` and `fluctuateListingPrefix`. These properties may be used to inject a computed class name on each listing item view (`fluctuateListingClass`).

This allows, for example, an alternating class to be applied to each item in the list (e.g. item-0 and item-1) so alternating rows can be styled distinctively.