Created big data helper with timer split for each processing
The Emberella.forEachAsync function iterates over an array for a specified amount of time. When time expires, the functions waits a bit, then proceeds and so on until all items in the array have been processed.

Helps to prevent the browser from locking up when performing operations on every member of an array.