Added 'highlighter' property to tags input
Set 'highlighter' to an object where each key is a class that may be added to the tag item and each value is a function.

Each function in the object will be called in the context of the tag item view. The content of the item view will be passed as an argument.

If the function returns a truthy value, then the associated object key will be added as a class on the tag item's element.

Otherwise, no additional classes will be added.

Use this functionality to alter the appearance of tags that match certain criteria.