Improved computation of "active" selection
The "active" selection is typically the intersection of the current selection set and the current content array. That is, an item only appears actively selected when it is both in the selection set and the content.

A new `getActiveContent` method for determining the content array from which active selections must be a member of was added to allow multi array controllers and the selectable mixin to work together.